reload tea choices when new tea is added
Currently this Flask instance only loads the tea choices in tea_log once, when the class is initiated. This means, functionally, that when a user adds a new tea and then tries to log it, that new tea choice does not appear in the dropdown menu of choices, since the class has not refreshed the list.

This commit moves that functionality to the route rather than the form (ie the class), so it is reloaded each time the user loads the page.